requirements.txt
# Core dependencies
streamlit==1.32.0
openai==1.10.0
python-dotenv==1.0.0
pydantic==2.4.2
flask==2.0.1
flask-restful==0.3.9
pinecone-client==2.2.1

# Database
sqlalchemy==2.0.23
sqlite3-api==2.0.4
redis==5.0.0

# Ad delivery and matching
scikit-learn==1.3.2
nltk==3.8.1
pandas==2.1.1
numpy==1.26.0

# Metrics and monitoring
prometheus-client==0.17.1
psutil==5.9.5

# Visualization
matplotlib==3.8.0
seaborn==0.13.0
plotly==4.14.0

# API and web
fastapi==0.104.1
uvicorn==0.23.2
requests==2.31.0
jinja2==3.1.2
urllib3==2.0.7

# Documentation
mkdocs==1.6.0
mkdocs-material==9.6.5

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1
rich==12.6.0

# Testing
pytest==8.0.0
pytest-cov==6.0.0
black==23.0.0
isort==5.0.0
flake8==7.0.0
mypy==1.0.0 